update staff details

+++ receipts/api.py
@@ -2,29 +2,46 @@ from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
 from rest_framework import status
 from django.shortcuts import get_object_or_404
-from .models import Receipt,AutoCorrectionLog
-from subscriptions.models import Subscription , SubscriptionType
+from .models import Receipt, AutoCorrectionLog
+from subscriptions.models import Subscription, SubscriptionType
 from members.models import Member
 from .serializers import ReceiptSerializer
 from rest_framework.permissions import IsAuthenticated
 from utils.permissions import IsOwnerOrRelatedToClub  
-from finance.models import Income , IncomeSource
+from finance.models import Income, IncomeSource
 from django.utils.timezone import now
 from rest_framework.pagination import PageNumberPagination
 from django.db.models import Q
 from django.utils import timezone
+from staff.models import StaffAttendance
+from datetime import datetime, timedelta
+
 
 @api_view(['GET'])
 @permission_classes([IsAuthenticated, IsOwnerOrRelatedToClub])
 def receipt_list_api(request):
     identifier = request.query_params.get('identifier', None)
     
-    if request.user.role == 'owner':
-        receipts = Receipt.objects.select_related('club', 'member', 'subscription', 'issued_by').all()
-    else:
+    if request.user.role in ['owner', 'admin']:
         receipts = Receipt.objects.select_related('club', 'member', 'subscription', 'issued_by').filter(club=request.user.club)
+    else:
+        attendance = StaffAttendance.objects.filter(
+            staff=request.user,
+            club=request.user.club
+        ).order_by('-check_in').first()
+
+        if not attendance:
+            return Response({'error': 'No open or recent shift found.'}, status=status.HTTP_404_NOT_FOUND)
+
+        check_out = attendance.check_out if attendance.check_out else timezone.now()
+        receipts = Receipt.objects.select_related('club', 'member', 'subscription', 'issued_by').filter(
+            club=request.user.club,
+            date__range=(attendance.check_in, check_out)
+        )
 
     if identifier:
+        if request.user.role not in ['owner', 'admin']:
+            receipts = Receipt.objects.select_related('club', 'member', 'subscription', 'issued_by').filter(club=request.user.club)
         receipts = receipts.filter(
             member__in=Member.objects.filter(
                 Q(name__icontains=identifier) |
@@ -32,6 +49,7 @@ def receipt_list_api(request):
                 Q(rfid_code__icontains=identifier)
             )
         )
+
     receipts = receipts.order_by('-date')
     paginator = PageNumberPagination()
     paginated_receipts = paginator.paginate_queryset(receipts, request)
@@ -40,7 +58,17 @@ def receipt_list_api(request):
 
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated, IsOwnerOrRelatedToClub])
 def add_receipt_api(request):
+    if request.user.role not in ['owner', 'admin']:
+        attendance = StaffAttendance.objects.filter(
+            staff=request.user,
+            club=request.user.club
+        ).order_by('-check_in').first()
+
+        if not attendance or (attendance.check_out and attendance.check_out < timezone.now()):
+            return Response({'error': 'You can only add receipts during an active shift.'}, status=status.HTTP_403_FORBIDDEN)
+
     data = request.data.copy()
     data['issued_by'] = request.user.id
 
@@ -85,48 +113,88 @@ def add_receipt_api(request):
     
     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+
 @api_view(['GET'])
 @permission_classes([IsAuthenticated, IsOwnerOrRelatedToClub])
 def receipt_detail_api(request, receipt_id):
     receipt = get_object_or_404(Receipt, id=receipt_id)
+    if request.user.role not in ['owner', 'admin']:
+        attendance = StaffAttendance.objects.filter(
+            staff=request.user,
+            club=request.user.club
+        ).order_by('-check_in').first()
+
+        if not attendance:
+            return Response({'error': 'No open or recent shift found.'}, status=status.HTTP_404_NOT_FOUND)
+
+        check_out = attendance.check_out if attendance.check_out else timezone.now()
+        if not (attendance.check_in <= receipt.date <= check_out):
+            return Response({'error': 'This receipt is not associated with your current shift.'}, status=status.HTTP_403_FORBIDDEN)
 
     serializer = ReceiptSerializer(receipt)
     return Response(serializer.data)
 
+
 @api_view(['PUT'])
 @permission_classes([IsAuthenticated, IsOwnerOrRelatedToClub])
 def edit_receipt_api(request, receipt_id):
-    receipt = get_object_or_404(Receipt, id=receipt_id)
+    if request.user.role not in ['owner', 'admin']:
+        return Response({'error': 'Only owners or admins can edit receipts.'}, status=status.HTTP_403_FORBIDDEN)
 
+    receipt = get_object_or_404(Receipt, id=receipt_id)
     serializer = ReceiptSerializer(receipt, data=request.data, partial=True)
     if serializer.is_valid():
         updated_receipt = serializer.save()
-
         if not IsOwnerOrRelatedToClub().has_object_permission(request, None, updated_receipt):
             return Response({'error': 'You do not have permission to update this receipt to this club'}, status=status.HTTP_403_FORBIDDEN)
         return Response(serializer.data)
     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+
 @api_view(['DELETE'])
 @permission_classes([IsAuthenticated, IsOwnerOrRelatedToClub])
 def delete_receipt_api(request, receipt_id):
-    receipt = get_object_or_404(Receipt, id=receipt_id)
+    if request.user.role not in ['owner', 'admin']:
+        return Response({'error': 'Only owners or admins can delete receipts.'}, status=status.HTTP_403_FORBIDDEN)
 
+    receipt = get_object_or_404(Receipt, id=receipt_id)
     receipt.delete()
     return Response(status=status.HTTP_204_NO_CONTENT)
+
 
 @api_view(['GET'])
 @permission_classes([IsAuthenticated, IsOwnerOrRelatedToClub])
 def receipt_by_invoice_api(request, invoice_number):
     receipt = get_object_or_404(Receipt, invoice_number=invoice_number)
+    if request.user.role not in ['owner', 'admin']:
+        attendance = StaffAttendance.objects.filter(
+            staff=request.user,
+            club=request.user.club
+        ).order_by('-check_in').first()
+
+        if not attendance:
+            return Response({'error': 'No open or recent shift found.'}, status=status.HTTP_404_NOT_FOUND)
+
+        check_out = attendance.check_out if attendance.check_out else timezone.now()
+        if not (attendance.check_in <= receipt.date <= check_out):
+            return Response({'error': 'This receipt is not associated with your current shift.'}, status=status.HTTP_403_FORBIDDEN)
 
     serializer = ReceiptSerializer(receipt)
     return Response(serializer.data)
 
-from datetime import datetime, timedelta
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated, IsOwnerOrRelatedToClub])
 def validate_or_autocorrect_receipt(request):
+    if request.user.role not in ['owner', 'admin']:
+        attendance = StaffAttendance.objects.filter(
+            staff=request.user,
+            club=request.user.club
+        ).order_by('-check_in').first()
+
+        if not attendance or (attendance.check_out and attendance.check_out < timezone.now()):
+            return Response({'error': 'You can only validate receipts during an active shift.'}, status=status.HTTP_403_FORBIDDEN)
+
     data = request.data
 
     member_id = data.get('member')
